requirements.txt
pyannote.audio
speechbrain
torchaudio
soundfile
huggingface_hub
python-dotenv
omegaconf
requests
dotenv
fastapi
uvicorn
pydantic
faster-whisper
openai-whisper
